DocQualityChecker/requirements.txt
python-docx
streamlit
openai
pytest
